+++ src/converters/resources/compute.ts
@@ -78,6 +78,8 @@ export function createDiskEntity(
     createdOn: getTime(data.timeCreated),
     webLink: webLinker.portalResourceUrl(data.id),
     encrypted: !!data.encryption?.type,
+    state: data.diskState?.toLowerCase(),
+    attached: data.diskState === "Attached",
   };
 
   assignTags(entity, data.tags);